Discovered and fixed bugs in CLI
I discovered that if you try adding a field or method to a non-existing class it will display that it did but in console show it did not. To fix this I added a classExist function in UMLClass so I could do a boolean check and see if the class exists already or not. I also added a list AllClassInfo command so you can see all the classes you made with their fields and methods.

+++ UML editor/src/main/java/codecain/CLI.java
@@ -169,9 +169,9 @@ public class CLI extends JFrame {
                 2. delete relationship 'source' 'destination' - Deletes the relationship between 'source' and 'destination'.
 
                 Field Operations:
-                1. add field 'className' 'fieldName'  - Adds a unique field to the specified class.
-                2. delete field 'className' 'fieldName' - Removes a field from the specified class.
-                3. rename field 'className' 'oldFieldName' 'newFieldName' - Renames a field in the specified class.
+                1. add field 'className' 'fieldName' 'fieldType'  - Adds a unique field to the specified class.
+                2. delete field 'className' 'fieldName''fieldType' - Removes a field from the specified class.
+                3. rename field 'className' 'oldFieldName' 'oldFieldType''newFieldName' 'newFieldType' - Renames a field and it's type in the specified class.
 
                 Method Operations:
                 1. add method 'className' 'methodName' 'parameters' - Adds a unique method to the specified class.
@@ -187,8 +187,9 @@ public class CLI extends JFrame {
 
                 Listing Operations:
                 1. list classes                        - Lists all the classes in the project.
-                2. list class 'className'              - Lists the contents (fields and methods) of the specified class.
+                2. list class 'className'              - (doesn't work) Lists the contents (fields and methods) of the specified class.
                 3. list relationships                  - Lists all the relationships between classes.
+                4. list AllClassInfo                   - List all the classes with their fields and methods.
 
                 Other Commands:
                 1. help                                - Shows this help message.
@@ -197,11 +198,16 @@ public class CLI extends JFrame {
                 Examples:
                 - add class Person
                 - add relationship Person Address
-                - add field Person name
+                
+                - add field Person name String
+                - rename field Person name String example int
+                
                 - add method Person setName String name
                 - add parameter Person setName String name
+                
                 - delete parameter Person setName String name
                 - rename parameter Person setName oldName newName
+                
                 - delete class Person
                 - rename class Person Employee
                 - list classes
@@ -235,28 +241,44 @@ public class CLI extends JFrame {
 
         String type = tokens[1].toLowerCase();
 
+        String className = tokens[2];
+
+
         switch (type) {
             case "class":
+
                 if (tokens.length == 3) {
                     return addClass(tokens[2]);
                 }
                 break;
             case "relationship":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 4) {
                     return addRelationship(tokens[2], tokens[3]);
                 }
                 break;
             case "field":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 5) {
                     return addField(tokens[2], tokens[3],tokens[4]);
                 }
                 break;
             case "method":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length >= 5) {
                     return addMethod(tokens[2], tokens[3], Arrays.copyOfRange(tokens, 4, tokens.length));
                 }
                 break;
             case "parameter":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 5) {
                     return addParameter(tokens[2], tokens[3], tokens[4]);
                 }
@@ -361,6 +383,7 @@ public class CLI extends JFrame {
         }
 
         String type = tokens[1].toLowerCase();
+        String className = tokens[2];
 
         switch (type) {
             case "class":
@@ -369,21 +392,33 @@ public class CLI extends JFrame {
                 }
                 break;
             case "relationship":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 4) {
                     return deleteRelationship(tokens[2], tokens[3]);
                 }
                 break;
             case "field":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 5) {
                     return deleteField(tokens[2], tokens[3],tokens[4]);
                 }
                 break;
             case "method":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 4) {
                     return deleteMethod(tokens[2], tokens[3]);
                 }
                 break;
             case "parameter":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 5) {
                     return deleteParameter(tokens[2], tokens[3], tokens[4]);
                 }
@@ -474,6 +509,7 @@ public class CLI extends JFrame {
         }
 
         String type = tokens[1].toLowerCase();
+        String className = tokens[2];
 
         switch (type) {
             case "class":
@@ -482,11 +518,17 @@ public class CLI extends JFrame {
                 }
                 break;
             case "field":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 7) {
                     return renameField(tokens[2], tokens[3], tokens[4], tokens[5],tokens[6]);
                 }
                 break;
             case "method":
+                if (!UMLClass.exists(className)) {
+                    return "Class '" + className + "' does not exist.";
+                }
                 if (tokens.length == 5) {
                     return renameMethod(tokens[2], tokens[3], tokens[4]);
                 }
@@ -549,6 +591,8 @@ public class CLI extends JFrame {
             return listClasses();
         } else if (tokens.length == 2 && tokens[1].equalsIgnoreCase("relationships")) {
             return listRelationships();
+        } else if (tokens.length == 2 && tokens[1].equalsIgnoreCase("AllClassInfo")) {
+            return UMLClass.listAllClassesInfo();
         } else {
             return "Invalid command. Use: list classes.";
         }
